# packages/api/openapi.yml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: ciel-api
  description: |
    Minimal SNS API (MVP).

    Auth is a simplified SCRAM-SHA-256 style challenge-response scheme.
    - Registration stores salted+iterated password verifier (storedKey/serverKey).
    - Login uses one-time challenge (nonce) with expiration to prevent replay.
  license:
    name: MIT
servers:
  - url: http://localhost:6137/api/v1
...
paths:
...
  /admin/media:
    get:
      tags: [Admin]
      summary: List media (admin view)
      description: List all media with filtering and admin-only fields
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by user ID
        - name: deleted
          in: query
          schema:
            type: boolean
          description: Filter by deletion status
      responses:
        '200':
          description: List of media
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMediaPage'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires admin_media_read permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /media:
    post:
      tags: [Media]
      summary: Upload media (image)
      description: |
        Upload an image and store it as WebP.

        **Allowed formats:** PNG, JPG, JPEG, WebP  
        **File size limit:** 12 MiB  
        **Input dimension limits:** 16384x16384 pixels, 100 megapixels total  
        **Output resolution:** Automatically resized to max 1920px (longest edge), aspect ratio preserved  
        
        **Security features:**
        - MIME type validation (content sniffing)
        - Metadata stripping (EXIF/XMP/GPS)
        - Automatic format conversion to WebP
        - DoS protection via size and dimension limits

        **Automatic resizing:**
        Images larger than 1920px on any edge are automatically resized to fit within 1920px
        while preserving the original aspect ratio. For example:
        - 6000x4000 → 1920x1280
        - 4000x6000 → 1280x1920
        - 4096x4096 → 1920x1920

        This means images that previously returned "image too large" errors (above 4096x4096)
        will now be accepted and automatically resized.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MediaId:
      type: string
      format: uuid
    MediaType:
      type: string
      enum: [image]
    Media:
      type: object
      required: [id, type, url, width, height, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/MediaId'
        type:
          $ref: '#/components/schemas/MediaType'
        url:
          type: string
          format: uri
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    AdminMedia:
      allOf:
        - $ref: '#/components/schemas/Media'
        - type: object
          properties:
            userId:
              type: string
              format: uuid
              description: User who uploaded this media
            uploaderUsername:
              type: string
              nullable: true
              description: Username of uploader
            deletedAt:
              type: string
              format: date-time
              nullable: true
              description: When the media was deleted
            deletedBy:
              type: string
              format: uuid
              nullable: true
              description: Admin user ID who deleted this media
            deletionReason:
              type: string
              nullable: true
              description: Reason for deletion
            phash:
              type: string
              nullable: true
              description: Perceptual hash of the image
            usedInPostsCount:
              type: integer
              description: Number of posts using this media
    AdminMediaPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminMedia'
        total:
          type: integer
          description: Total number of media items matching the filters
...